+++ src/buildTx.ts
@@ -45,6 +45,12 @@ export const buildTx = async ({
   let coinData: TransactionResult;
   if (coinIn) {
     coinData = tx.splitCoins(coinIn, splits);
+    SuiUtils.transferOrDestroyZeroCoin(
+      tx,
+      quoteResponse.tokenIn,
+      coinIn,
+      accountAddress,
+    );
   } else {
     const { coinData: _data } = await getSplitCoinForTx(
       accountAddress,